refactor: re-paper crt engine to smithy kotlin and incorporate into common test suite
