+++ gradle.properties
@@ -39,3 +39,6 @@ kotlinLoggingVersion=2.0.3
 
 # logging - JVM
 slf4jVersion=1.7.30
+
+# crt
+crtKotlinVersion=0.5.4-SNAPSHOT
