+++ runtime/crt-util/build.gradle.kts
@@ -17,34 +17,31 @@ buildscript {
 apply(plugin = "kotlinx-atomicfu")
 
 description = "Utilities for working with AWS CRT Kotlin"
-extra["displayName"] = "AWS :: SDK :: Kotlin :: CRT :: Util"
-extra["moduleName"] = "aws.sdk.kotlin.runtime.crt"
+extra["displayName"] = "Smithy :: Kotlin :: CRT :: Util"
+extra["moduleName"] = "aws.smithy.kotlin.runtime.crt"
 
 val atomicFuVersion: String by project
 val coroutinesVersion: String by project
 val crtKotlinVersion: String by project
-val smithyKotlinVersion: String by project
 
 kotlin {
     sourceSets {
         commonMain {
             dependencies {
-                api(project(":aws-runtime:aws-core"))
+                api(project(":runtime:runtime-core"))
                 api("aws.sdk.kotlin.crt:aws-crt-kotlin:$crtKotlinVersion")
-                api("aws.smithy.kotlin:http:$smithyKotlinVersion")
+                api(project(":runtime:protocol:http"))
                 implementation("org.jetbrains.kotlinx:atomicfu:$atomicFuVersion")
             }
         }
         commonTest {
             dependencies {
-                implementation(project(":aws-runtime:testing"))
                 implementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:$coroutinesVersion")
             }
         }
 
         all {
             languageSettings.optIn("aws.smithy.kotlin.runtime.util.InternalApi")
-            languageSettings.optIn("aws.sdk.kotlin.runtime.InternalSdkApi")
         }
     }
 }
