+++ runtime/protocol/http-client-engines/test-suite/jvm/src/aws/smithy/kotlin/runtime/http/test/util/AbstractEngineTestJVM.kt
@@ -6,6 +6,7 @@
 package aws.smithy.kotlin.runtime.http.test.util
 
 import aws.smithy.kotlin.runtime.http.engine.HttpClientEngine
+import aws.smithy.kotlin.runtime.http.engine.crt.CrtHttpEngine
 import aws.smithy.kotlin.runtime.http.engine.ktor.KtorEngine
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.runBlocking
@@ -15,7 +16,8 @@ import kotlin.time.Duration
 
 internal actual fun engines(): List<HttpClientEngine> =
     listOf(
-        KtorEngine()
+        KtorEngine(),
+        CrtHttpEngine()
     )
 
 internal actual fun runBlockingTest(
