+++ runtime/runtime-core/build.gradle.kts
@@ -9,7 +9,7 @@ plugins {
 
 description = "Client runtime for Smithy services generated by smithy-kotlin"
 extra["displayName"] = "Smithy :: Kotlin :: Client Runtime"
-extra["moduleName"] = "software.amazon.smithy.kotlin.runtime"
+extra["moduleName"] = "aws.smithy.kotlin.runtime"
 
 val coroutinesVersion: String by project
