+++ settings.gradle.kts
@@ -40,9 +40,11 @@ include(":runtime:serde")
 include(":runtime:serde:serde-json")
 include(":runtime:serde:serde-xml")
 include(":runtime:serde:serde-form-url")
+include(":runtime:crt-util")
 include(":runtime:protocol:http")
 include(":runtime:protocol:http-test")
 include(":runtime:protocol:http-client-engines:http-client-engine-ktor")
+include(":runtime:protocol:http-client-engines:http-client-engine-crt")
 include(":runtime:protocol:http-client-engines:test-suite")
 
 include(":tests")
